LayoutTests/fast/events/offsetX-offsetY-svg-expected.txt


PASS MouseEvent.offsetX/offsetY for an SVG element is always relative to the outermost <svg> CSS layout box
